container-applications/IoT/cloud_function/requirements.txt
click
dominate
Flask
itsdangerous
Jinja2
MarkupSafe
onetimepass
six
visitor
pymysql
Werkzeug
gunicorn
google-api-core
google-cloud-datastore
google-api-python-client
google-cloud-core
google-cloud-storage
google-cloud-runtimeconfig
google-cloud-pubsub
google-cloud-iot
requests
pytz
google-cloud-error-reporting==1.9.1
google-cloud-logging